pyproject.toml
[tool.poetry]
name = "python-prj-sample"
version = "0.1.0"
description = ""
authors = ["hoge <hoge@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = {extras = ["cli"], version = "^0.19.0"}
fastapi = "^0.75.1"
boto3 = "^1.21.35"
uvicorn = "^0.17.6"
python-multipart = "^0.0.5"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pysen = "0.10.2"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
pytest-sugar = "^0.9.4"
botostubs = "^0.15.1"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.950"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10.2"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 120
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["src"]
